+++ src/app/(sikundi)/sikundi-admin/(collections)/video/_actions/update/index.ts
@@ -5,6 +5,7 @@ import VideoSchema, { VideoSchemaType } from './schema'
 import ErrorHandler from '@sikundi/lib/server/utils/ErrorHandler'
 import { prisma } from '@sikundi/lib/server/utils/prisma'
 import { revalidatePath } from 'next/cache'
+import getPermission from '@sikundi/lib/server/utils/getPermission'
 
 const statusFromActions = {
     draft: "drafted",
@@ -16,6 +17,25 @@ const statusFromActions = {
 export default async function POST(data: VideoSchemaType) {
     return (await ErrorHandler(data, VideoSchema, async ({createdBy, language, push, action, id, ...data}:VideoSchemaType) => {
         const user = await getUser()
+        const permission = await getPermission({
+            video: [
+                "draft",
+                "delete",
+                "soft_delete",
+                "publish",
+                "pending"
+            ]
+        })
+
+        if(!permission?.video?.[String(action)]) {
+            throw({
+                notification: {
+                    title: 'Authorization Error',
+                    description: `You are not allowed to ${action} videos.`,
+                    variant: "destructive"
+                }
+            })
+        }
 
         if (action === "delete") {
             const video = await prisma.video.delete({
